# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Server Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type BotInfoStatus = (typeof BotInfoStatus)[keyof typeof BotInfoStatus];

export const BotInfoStatus = {
  online: "online",
  idle: "idle",
  dnd: "dnd",
  offline: "offline",
} as const;

export interface BotInfo {
  id: string;
  username: string;
  discriminator: string;
  avatar?: string | null;
  banner?: string | null;
  bio?: string | null;
  guildCount: number;
  status: BotInfoStatus;
  ping: number;
}

export interface Guild {
  id: string;
  name: string;
  icon?: string | null;
  memberCount: number;
  ownerId: string;
}

export interface GuildOverview {
  guildId: string;
  memberCount: number;
  onlineCount: number;
  channelCount: number;
  roleCount: number;
  boostLevel: number;
  boostCount: number;
  warningsToday: number;
  bansToday: number;
  messagesProcessed: number;
  antinukeEnabled: boolean;
  antiraidEnabled: boolean;
}

export type ChannelType = (typeof ChannelType)[keyof typeof ChannelType];

export const ChannelType = {
  text: "text",
  voice: "voice",
  category: "category",
  announcement: "announcement",
  forum: "forum",
  stage: "stage",
} as const;

export interface Channel {
  id: string;
  name: string;
  type: ChannelType;
}

export interface Role {
  id: string;
  name: string;
  color: number;
  position: number;
}

/**
 * Action to take against the rogue user
 */
export type AntinukeSettingsAction =
  (typeof AntinukeSettingsAction)[keyof typeof AntinukeSettingsAction];

export const AntinukeSettingsAction = {
  ban: "ban",
  kick: "kick",
  strip_roles: "strip_roles",
  timeout: "timeout",
} as const;

export interface AntinukeSettings {
  enabled: boolean;
  /** Max channel deletes per interval before action */
  maxChannelDeletes: number;
  /** Max bans per interval before action */
  maxBans: number;
  /** Max kicks per interval before action */
  maxKicks: number;
  maxRoleDeletes: number;
  maxWebhookCreates: number;
  /** Time window in seconds to count actions */
  intervalSeconds: number;
  /** Action to take against the rogue user */
  action: AntinukeSettingsAction;
  logChannelId?: string | null;
  dmOwner: boolean;
}

export type WhitelistEntryTargetType =
  (typeof WhitelistEntryTargetType)[keyof typeof WhitelistEntryTargetType];

export const WhitelistEntryTargetType = {
  user: "user",
  role: "role",
} as const;

export interface WhitelistEntry {
  id?: number;
  guildId?: string;
  /** User or role ID */
  targetId: string;
  targetType: WhitelistEntryTargetType;
  targetName: string;
  addedAt?: string;
}

export type AntiraidSettingsAction =
  (typeof AntiraidSettingsAction)[keyof typeof AntiraidSettingsAction];

export const AntiraidSettingsAction = {
  kick: "kick",
  ban: "ban",
  timeout: "timeout",
  lock: "lock",
} as const;

export interface AntiraidSettings {
  enabled: boolean;
  /** Max joins per interval */
  joinRateLimit: number;
  /** Interval in seconds */
  joinRateInterval: number;
  action: AntiraidSettingsAction;
  /** Kick accounts without avatars */
  filterNoAvatar: boolean;
  /** Kick accounts newer than minAccountAge */
  filterNewAccounts: boolean;
  minAccountAgeDays: number;
  lockdownEnabled: boolean;
  logChannelId?: string | null;
}

export type AutomodSettingsSpamAction =
  (typeof AutomodSettingsSpamAction)[keyof typeof AutomodSettingsSpamAction];

export const AutomodSettingsSpamAction = {
  delete: "delete",
  warn: "warn",
  mute: "mute",
  kick: "kick",
  ban: "ban",
} as const;

export interface AutomodSettings {
  enabled: boolean;
  filterInvites: boolean;
  filterLinks: boolean;
  filterSpam: boolean;
  filterCaps: boolean;
  /** Percentage of caps to trigger filter */
  capsThreshold: number;
  filterMentionSpam: boolean;
  maxMentions: number;
  filterWords: boolean;
  bannedWords: string[];
  spamAction: AutomodSettingsSpamAction;
  logChannelId?: string | null;
}

export interface JailSettings {
  enabled: boolean;
  jailChannelId?: string | null;
  jailRoleId?: string | null;
  logChannelId?: string | null;
}

export interface LevelRole {
  level: number;
  roleId: string;
  roleName: string;
}

export interface LevelingSettings {
  enabled: boolean;
  xpPerMessage: number;
  xpCooldownSeconds: number;
  levelUpMessage: string;
  levelUpChannelId?: string | null;
  ignoredChannels: string[];
  ignoredRoles: string[];
  stackRoles: boolean;
  levelRoles: LevelRole[];
}

export interface ReactionRole {
  id?: number;
  guildId?: string;
  channelId: string;
  messageId: string;
  emoji: string;
  roleId: string;
  roleName: string;
  description?: string | null;
}

export interface WelcomeSettings {
  welcomeEnabled: boolean;
  welcomeChannelId?: string | null;
  welcomeMessage: string;
  welcomeEmbed: boolean;
  goodbyeEnabled: boolean;
  goodbyeChannelId?: string | null;
  goodbyeMessage: string;
  dmWelcome: boolean;
  dmMessage: string;
}

export type SocialAlertPlatform =
  (typeof SocialAlertPlatform)[keyof typeof SocialAlertPlatform];

export const SocialAlertPlatform = {
  twitter: "twitter",
  tiktok: "tiktok",
  instagram: "instagram",
  youtube: "youtube",
  twitch: "twitch",
} as const;

export interface SocialAlert {
  id?: number;
  guildId?: string;
  platform: SocialAlertPlatform;
  accountHandle: string;
  channelId: string;
  message: string;
  enabled: boolean;
  lastChecked?: string | null;
}

export type AuditLogType = (typeof AuditLogType)[keyof typeof AuditLogType];

export const AuditLogType = {
  antinuke_triggered: "antinuke_triggered",
  antiraid_triggered: "antiraid_triggered",
  automod_action: "automod_action",
  member_ban: "member_ban",
  member_kick: "member_kick",
  member_warn: "member_warn",
  member_jail: "member_jail",
  member_unjail: "member_unjail",
  settings_change: "settings_change",
  bot_action: "bot_action",
} as const;

export interface AuditLog {
  id: number;
  guildId: string;
  type: AuditLogType;
  executorId?: string | null;
  executorName?: string | null;
  targetId?: string | null;
  targetName?: string | null;
  reason?: string | null;
  details?: string | null;
  createdAt: string;
}

export type GetAuditLogsParams = {
  limit?: number;
  type?: string;
};