# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Server Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Role {
  id: string;
  name: string;
  color: number;
  position: number;
}