# lib/api-zod/src/generated/types/welcomeSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Server Management API
 * OpenAPI spec version: 0.1.0
 */

export interface WelcomeSettings {
  welcomeEnabled: boolean;
  welcomeChannelId?: string | null;
  welcomeMessage: string;
  welcomeEmbed: boolean;
  goodbyeEnabled: boolean;
  goodbyeChannelId?: string | null;
  goodbyeMessage: string;
  dmWelcome: boolean;
  dmMessage: string;
}